visualize

+++ demo_patch.py
@@ -216,7 +216,7 @@ for id, path in enumerate(fpaths):
     filename = path.split('/')[-2]
     save_name = os.path.join( ADV_IMAGE_FOLDER, filename)
     print(x_adv.shape)
-    img = Image.fromarray(x_adv[0].int())
+    img = Image.fromarray(x_adv[0])
     print(img)
     # adv_image = x_adv.SAVE(save_name)
